A-739: The time slots should not be far in the future (#300)
* A-739: Add check to verify that blocks come from the past in `isValid` method

Changes
- Add time slot parameter to validation methods of `SegmentValidator`
- Check that block comes from the past in block validation method
- Add slot duration parameter to `Server` class
- Add slot duration parameter to `Blockchain` class
- Change `TimeSlot` underlaying representation to `Long`
- Change `H2BlockStorage` table definition to update `time_slot` column to use `LONG` type
- Update tests

* A-739: Address PR reviews

Changes:
- Remove slot duration from `Blockchain` class
- Add slot duration to `SegmentValidator` class
- Accept blocks emitted in the current time slot

+++ atala/obft/blockchain/storage/H2BlockStorage.scala
@@ -136,7 +136,7 @@ object H2BlockStorage {
   }
 
   private implicit val timeSlotGet: Get[TimeSlot] =
-    Get[Int].tmap(x => TimeSlot.from(x).getOrElse(throw new RuntimeException("Corrupted timeslot")))
+    Get[Long].tmap(x => TimeSlot.from(x).getOrElse(throw new RuntimeException("Corrupted timeslot")))
 
   private implicit def deltaGet[Tx](implicit txCodec: Codec[List[Tx]]): Read[List[Tx]] = Read[Array[Byte]].map {
     bytes =>
@@ -163,7 +163,7 @@ object H2BlockStorage {
          |  delta BINARY NOT NULL,
          |  previous_hash VARCHAR NOT NULL,
          |  signature VARCHAR NOT NULL,
-         |  time_slot INT NOT NULL UNIQUE,
+         |  time_slot LONG NOT NULL UNIQUE,
          |  time_slot_signature VARCHAR NOT NULL
          |);
          """.stripMargin
